Wayland: Add initial support for libdecor
This is partly based on the implementation of libdecor support in
PR #1693 by @ christianrauch.

The libdecor library is loaded and preferred where available.  On
compositors that support XDG decorations, libdecor uses those.  It in
turn has a plug-in archtecture and may load additional libraries to use
compositor-specific decorations or draw its own.

If necessary, support for libdecor can be disabled with the
GLFW_WAYLAND_LIBDECOR init hint.  This is in case some part of the
dynamic loading or duplication of header material I added here turns out
to cause problems with future versions of libdecor-0.so.0.

Fixes #1639
Closes #1693